Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,19 +1,19 @@
 #include <stdio.h>
 int main(){
-    float a,b;
+    int a,b;
     char ch;
     scanf("%d %d %ch",&a,&b,&ch);
     if(ch=='+'){
-        printf("%f",a+b);
+        printf("%i",a+b);
     }
     else if(ch=='-'){
-        printf("%f",a-b);
+        printf("%i",a-b);
     }
     else if(ch=='*'){
-        printf("%f",a*b);
+        printf("%i",a*b);
     }
     else if(ch=='/'){
-        printf("%f",a/b);
+        printf("%i",a/b);
     }
     else{
         printf("error");
